chore: move some plugins
